+++ fabric/changelog.md
@@ -1,1 +1,2 @@
 - Fixed group not getting deleted when disconnecting
+- Fixed group in LeaveGroupEvent always null